fix image control zooming every frame
