+++ Dat/Loaders/TrackObjectLoader.cs
@@ -1,10 +1,8 @@
 using Dat.Data;
 using Dat.FileParsing;
-using Dat.Types;
 using Definitions.ObjectModels;
 using Definitions.ObjectModels.Objects.Track;
 using Definitions.ObjectModels.Types;
-using System.ComponentModel;
 using static Dat.Loaders.TrackObjectLoader;
 
 namespace Dat.Loaders;
