+++ Gui/Views/ImageTableView.axaml
@@ -8,6 +8,7 @@
 	xmlns:pgc="clr-namespace:Avalonia.PropertyGrid.Controls;assembly=Avalonia.PropertyGrid"
 	xmlns:amxc="using:Avalonia.Markup.Xaml.Converters"
 	xmlns:gmc="using:Gui.Models.Converters"
+	xmlns:paz="using:Avalonia.Controls.PanAndZoom"
 	xmlns:vi="using:Gui.Views"
 	mc:Ignorable="d"
 	d:DesignWidth="1200"
@@ -17,6 +18,7 @@
 
 	<UserControl.Resources>
 		<amxc:ColorToBrushConverter x:Key="ColorToBrushConverter"/>
+		<gmc:NegativeValueConverter x:Key="NegativeValueConverter"/>
 	</UserControl.Resources>
 
 	<DockPanel>
@@ -94,6 +96,52 @@
 				<StackPanel Margin="2" Orientation="Vertical" MaxHeight="512" DockPanel.Dock="Top">
 					<TextBlock Margin="0, 16, 0, 0" HorizontalAlignment="Center" Text="{Binding AnimationSpeed, StringFormat='Animation FPS: {0}'}"></TextBlock>
 					<Slider Minimum="1" Maximum="80" Value="{Binding AnimationSpeed}" />
+					<StackPanel Orientation="Horizontal">
+						<Button HorizontalAlignment="Stretch" Margin="2" Padding="2" ToolTip.Tip="Select preview background colour">
+							<DockPanel>
+								<materialIcons:MaterialIcon Kind="Palette" Width="24" Height="24" Margin="2" />
+								<TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="4">Background</TextBlock>
+							</DockPanel>
+							<Button.Flyout >
+								<Flyout>
+									<ColorView Name="ImageColorViewPreviewBackground" Palette="" Color="SteelBlue" IsColorModelVisible="False" IsColorSpectrumVisible="False" IsColorComponentsVisible="False" />
+								</Flyout>
+							</Button.Flyout>
+						</Button>
+						<Button HorizontalAlignment="Stretch" Margin="2" Padding="2" ToolTip.Tip="Select border preview colour">
+							<DockPanel>
+								<materialIcons:MaterialIcon Kind="Palette" Width="24" Height="24" Margin="2" />
+								<TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="4">Border</TextBlock>
+							</DockPanel>
+							<Button.Flyout >
+								<Flyout>
+									<ColorView Name="ImageColorViewBorder" Palette="" Color="Gold" IsColorModelVisible="False" IsColorSpectrumVisible="False" IsColorComponentsVisible="False" />
+								</Flyout>
+							</Button.Flyout>
+						</Button>
+						<Button HorizontalAlignment="Stretch" Margin="2" Padding="2" ToolTip.Tip="Select origin point colour">
+							<DockPanel>
+								<materialIcons:MaterialIcon Kind="Palette" Width="24" Height="24" Margin="2" />
+								<TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="4">Origin</TextBlock>
+							</DockPanel>
+							<Button.Flyout >
+								<Flyout>
+									<ColorView Name="ImageColorViewOriginPoint" Palette="" Color="Red" IsColorModelVisible="False" IsColorSpectrumVisible="False" IsColorComponentsVisible="False" />
+								</Flyout>
+							</Button.Flyout>
+						</Button>
+					</StackPanel>
+					<paz:ZoomBorder Name="ZoomBorder" MinWidth="256" MinHeight="256" Stretch="None" ZoomX="{Binding ZoomLevel}" ZoomSpeed="1.2" MinZoomX="0.5" MaxZoomX="30" MinZoomY="0.5" MaxZoomY="30" ClipToBounds="True" Focusable="True" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" PanButton="Left"  Background="{Binding #ImageColorViewPreviewBackground.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}" >
+						<Canvas Width="{Binding SelectedImage.Width}" Height="{Binding SelectedImage.Height}" HorizontalAlignment="Center" VerticalAlignment="Center" Background="Gray">
+							<!--Bounding box-->
+							<Rectangle Canvas.Top="-1" Canvas.Left="-1" StrokeThickness="1.0" Width="{Binding SelectedImage.SelectedBitmapPreviewBorder.Width}" Height="{Binding SelectedImage.SelectedBitmapPreviewBorder.Height}" Opacity="1.0" Fill="{Binding #ImageColorViewPreviewBackground.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}" Stroke="{Binding #ImageColorViewBorder.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}"  />
+							<!--Image-->
+							<Image Name="AnimationPreviewPP" Canvas.Top="0" Canvas.Left ="0" Source="{Binding SelectedImage.Image}" RenderOptions.BitmapInterpolationMode="None" Stretch="None"/>
+							<!--Origin point-->
+							<Rectangle Canvas.Left="{Binding SelectedImage.XOffset, Converter={StaticResource NegativeValueConverter}}" Canvas.Top="{Binding SelectedImage.YOffset, Converter={StaticResource NegativeValueConverter}}" Width="1" Height="1" Opacity="1.0" Fill="{Binding #ImageColorViewOriginPoint.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}" />
+						</Canvas>
+					</paz:ZoomBorder>
+					<TextBlock HorizontalAlignment="Center" Text="{Binding #ZoomBorder.ZoomX, StringFormat='Zoom: {0:F2}x | \'R\' to reset'}"></TextBlock>
 				</StackPanel>
 					<ContentControl Content="{Binding SelectedImage}" />
 			</DockPanel>
