+++ Gui/Views/ImageView.axaml
@@ -6,70 +6,17 @@
 	xmlns:vm="using:Gui.ViewModels"
 	xmlns:materialIcons="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
 	xmlns:pgc="clr-namespace:Avalonia.PropertyGrid.Controls;assembly=Avalonia.PropertyGrid"
-	xmlns:paz="using:Avalonia.Controls.PanAndZoom"
+	
 	xmlns:amxc="using:Avalonia.Markup.Xaml.Converters"
 	xmlns:gmc="using:Gui.Models.Converters"
 	mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
 	x:Class="Gui.Views.ImageView"
 	x:DataType="vm:ImageViewModel">
 
-	<UserControl.Resources>
-		<amxc:ColorToBrushConverter x:Key="ColorToBrushConverter"/>
-		<gmc:NegativeValueConverter x:Key="NegativeValueConverter"/>
-	</UserControl.Resources>
-
 	<ScrollViewer Margin="4" Name="SV">
 		<DockPanel>
 			<Panel Width="16" DockPanel.Dock="Right"></Panel>
-			<StackPanel Orientation="Vertical">
-				<StackPanel Orientation="Horizontal">
-					<Button HorizontalAlignment="Stretch" Margin="2" Padding="2" ToolTip.Tip="Select preview background colour">
-						<DockPanel>
-							<materialIcons:MaterialIcon Kind="Palette" Width="24" Height="24" Margin="2" />
-							<TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="4">Background</TextBlock>
-						</DockPanel>
-						<Button.Flyout >
-							<Flyout>
-								<ColorView Name="ImageColorViewPreviewBackground" Palette="" Color="SteelBlue" IsColorModelVisible="False" IsColorSpectrumVisible="False" IsColorComponentsVisible="False" />
-							</Flyout>
-						</Button.Flyout>
-					</Button>
-					<Button HorizontalAlignment="Stretch" Margin="2" Padding="2" ToolTip.Tip="Select border preview colour">
-						<DockPanel>
-							<materialIcons:MaterialIcon Kind="Palette" Width="24" Height="24" Margin="2" />
-							<TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="4">Border</TextBlock>
-						</DockPanel>
-						<Button.Flyout >
-							<Flyout>
-								<ColorView Name="ImageColorViewBorder" Palette="" Color="Gold" IsColorModelVisible="False" IsColorSpectrumVisible="False" IsColorComponentsVisible="False" />
-							</Flyout>
-						</Button.Flyout>
-					</Button>
-					<Button HorizontalAlignment="Stretch" Margin="2" Padding="2" ToolTip.Tip="Select origin point colour">
-						<DockPanel>
-							<materialIcons:MaterialIcon Kind="Palette" Width="24" Height="24" Margin="2" />
-							<TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="4">Origin</TextBlock>
-						</DockPanel>
-						<Button.Flyout >
-							<Flyout>
-								<ColorView Name="ImageColorViewOriginPoint" Palette="" Color="Red" IsColorModelVisible="False" IsColorSpectrumVisible="False" IsColorComponentsVisible="False" />
-							</Flyout>
-						</Button.Flyout>
-					</Button>
-				</StackPanel>
-				<paz:ZoomBorder Name="ZoomBorder" MinWidth="256" MinHeight="256" Stretch="Uniform" ZoomSpeed="1.2" MinZoomX="0.5" MaxZoomX="30" MinZoomY="0.5" MaxZoomY="30" ClipToBounds="True" Focusable="True" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" PanButton="Left"  Background="{Binding #ImageColorViewPreviewBackground.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}" >
-					<Canvas Width="{Binding Width}" Height="{Binding Height}" HorizontalAlignment="Center" VerticalAlignment="Center" Background="Gray">
-						<!--Bounding box-->
-						<Rectangle Canvas.Top="-1" Canvas.Left="-1" StrokeThickness="1.0" Width="{Binding SelectedBitmapPreviewBorder.Width}" Height="{Binding SelectedBitmapPreviewBorder.Height}" Opacity="1.0" Fill="{Binding #ImageColorViewPreviewBackground.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}" Stroke="{Binding #ImageColorViewBorder.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}"  />
-						<!--Image-->
-						<Image Name="AnimationPreviewPP" Canvas.Top="0" Canvas.Left ="0" Source="{Binding Image}" RenderOptions.BitmapInterpolationMode="None" Stretch="None"/>
-						<!--Origin point-->
-						<Rectangle Canvas.Left="{Binding XOffset, Converter={StaticResource NegativeValueConverter}}" Canvas.Top="{Binding YOffset, Converter={StaticResource NegativeValueConverter}}" Width="1" Height="1" Opacity="1.0" Fill="{Binding #ImageColorViewOriginPoint.Color, ConverterParameter={x:Static Brushes.Transparent}, Converter={StaticResource ColorToBrushConverter}}" />
-					</Canvas>
-				</paz:ZoomBorder>
-				<TextBlock HorizontalAlignment="Center" Text="{Binding #ZoomBorder.ZoomX, StringFormat='Zoom: {0:F2}x | \'R\' to reset'}"></TextBlock>
-				<pgc:PropertyGrid DataContext="{Binding}" IsCategoryVisible="False"></pgc:PropertyGrid>
-			</StackPanel>
+			<pgc:PropertyGrid DataContext="{Binding}" IsCategoryVisible="False"></pgc:PropertyGrid>
 		</DockPanel>
 	</ScrollViewer>
